Fix some markup that had a duplicate closing tag in the header. Add the javascript value the search-substitute code looks for.
Instead of including the content wrapper HTML in every template, we're going to include it in the main template, so add the two tags to header and close them off in footer.

Since we don't have room in every individual template now, load the sidebar in the footer as well.